Revert "Amount should be a double"
This reverts commit cc77c2d60a75cdd5dc68d61c4492c69d91197d15.

+++ src/main/java/com/MultiSafepay/client/MultiSafepayClient.java
@@ -65,7 +65,7 @@ public class MultiSafepayClient {
 		return sendRequest("orders/" + order_id +"/transactions");
 	}
 	
-	public JsonObject SetOrderRefund(String order_id,Double amount,String currency,String description) {
+	public JsonObject SetOrderRefund(String order_id,Integer amount,String currency,String description) {
 		Order order				= new Order();
     	order.currency			= currency;
     	order.amount			= amount;
